pyproject.toml
[tool.poetry]
name = "event-management-backend"
version = "0.1.0"
description = ""
authors = ["Rishav Sharma <rishishah907@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2.6"
djangorestframework = "^3.16.1"
drf-spectacular = "^0.28.0"
pytest = "^8.4.2"
pytest-django = "^4.11.1"
psycopg2-binary = "^2.9.10"
pre-commit = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
